tor-hsservice: replay test: Use two-rusty-forks
I'm using the forked crate because 1. the proc macro is nice 2. it
supports async, which we don't necessarily want now but may in the
future.

Unfortunately, this seems to make --nocapture not work.  I haven't
investigated this in detail, but the docs for rusty-fork do say some
stuff about capturing output.  The previous open-coded implementation
was rather simpler, and doesn't have this deficiency.

So I'm now in two minds whether this is a good idea.

If we accept this regression, we might want to consider reproing
the --nocapture bug with the original crate and filing an issue.