proto: Add ChannelType enum
The ChannelType indicates the type of channel in order to dictate which
message is allowed on it. The value use the Initiator and Responder
terminology from tor-spec documents.

At this commit, we only have client channel meaning the
"ClientInitiator" type.

In future commits, the channel type will be used by the channel reactor
to restrict which message is allowed or not.

Part of #1597

Signed-off-by: David Goulet <dgoulet@torproject.org>