proto: Add restricted channel message sets
Add the msg.rs file containing all the allowed message sets based on the
channel type and direction. They are also namespaced by link protocol
version.

Unused at this commit. They will be used by the channel reactor along
the channel type and link protocol version in order to know if the
message is allowed or not. See is_allowed() helper function in this
commit.

Part of #1597

Signed-off-by: David Goulet <dgoulet@torproject.org>