proto: Add message filtering to restricted message sets
This adds the code in msg.rs to be able to filter an inbound or outbound
message on a channel.

Each link protocol version implement a "is_allowed()" which is quite
verbose and tests each possibilities for human readability.

Then, we have several small struct/enum that are used to describe how a
message is filtered.

It is still unused at this commit.

Part of #1597

Signed-off-by: David Goulet <dgoulet@torproject.org>