tor-proto: Return EOF for all EndReceived errors
Originally the `AsyncRead` returned EOF only once and then kept returning errors. This resulted in clients not interpretting the EOF correctly and consistenly gives the error: "Error: Received an END cell with reason MISC". This issue occurred especially when setting up a reverse proxy using tor-hsrproxy.

This commit decides that it seems reasonable to return the EOF to the client for all cases of `EndReceived`. A less strict possibility would be to only return EOF for `EndReceived::DONE`, `EndReceived::MISC` and `EndReceived::DESTROY`.

Anyhow, we want to return EOF for more cases then only `EndReceived::DONE`.