tor-hsrproxy: Close both streams when one finishes
DataStream and one of the readers in `copy_interactive` consistently received the error
`Custom { kind: InvalidData, error: StreamProto("stream channel disappeared without END cell?")` when a request finished.

This because when one direction of the proxy finished, the other direction still tries to read from the stream which does not exists anymore.

In this commit we use `select!` to close the other direction when one is finished.