squash! arti console logging: use stderr instead of stdout
Don't document that setting logging.console to the empty string disables
it, since that doesn't seem to be the current behavior, and I'm not 100%
sure that it's the intended behavior.

I'll leave documenting how to disable console logging to if and when
there is a way to do so, as part of fixing #2025.