CI: increase CHUTNEY_START_TIME to default
This is the time we wait for the network to bootstrap before giving up.
With the larger network we seem to need this extra time.

We've already increased the default in chutney to 300; no reason to use
a smaller value here.

Keep the explicit-setting for now as a convenience in case we want to
modify it again.